requirements.txt
requests
python-dotenv
pydantic
pydantic-settings
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
gspread
gspread-formatting
aiohttp